UPBGE: Refactor profiler access from python.
Previously the profiler dict storing category and time was
generated each frame, but if the user doesn't used it was waste
time.
To avoid such situation, the timer logger can return a map of
the categories name and their last time via GetProfileDict().
This std::map is simply converted in KX_PythonInit gPyGetProfileInfo.

Notice that name categories are now "Name" and not odd "Name:".