temp
I'm not sure, but maybe this AA is a kindof optic effect. When you set
eevee_effect->overide matrices then when you redraw the materials, and
when you set the fps to 1 per second, you see the pixels at the edge of
a model moving. When you increase fps, the pixels at the edge move very
fast and this creates a kind of blur at the edge.

I don't understand all for now and this is not working well for now in
bge but at least we can see something happening on the edges.

This is WIP waiting we understand how it works more exactly...

TODO: See what is done with taa_current_sample in eevee_engine.c