UPBGE: Implement special linked list.
A new common class named CM_LinkedList is implemented to fill the
following goal:
Manage adding/removing of element from Link and check if the link
is added or not.

This look like SG_Q/DList but it is more simpler, not circular and
use template parameter to avoid messing Link and List types.

The link class is CM_LinkedList<>::Link, it offers three functions:
Add(), Remove(), IsRemoved().
To create a link, CM_LinkedList<>::NewLink must be called, by default
the link isn't added in the list.